+++ packages/block/src/custom.ts
@@ -40,13 +40,11 @@ class CustomBlockType extends BlockType {
 	 * @param loggable Whether the block type is loggable.
 	 */
 	public constructor(identifier: string, loggable: boolean) {
+		// TODO: custom states, hardness, friction, mapColor, components, tags, etc.
 		super(identifier as keyof BlockState, loggable, false, false, true);
 
 		// Construct the NBT tag.
 		this.nbt = new CompoundTag("", {});
-
-		// Register the block type.
-		BlockType.types.set(identifier, this);
 	}
 
 	/**