+++ packages/serenity/src/handlers/resource-pack-client-response.ts
@@ -13,12 +13,13 @@ import {
 	DisconnectReason,
 	type BlockProperties,
 	ResourceIdVersions,
-	CraftingDataPacket
+	CraftingDataPacket,
+	AvailableActorIdentifiersPacket
 } from "@serenityjs/protocol";
 import { BIOME_DEFINITION_LIST, CRAFTING_DATA } from "@serenityjs/data";
 import { CreativeItem, CustomItemType, ItemType } from "@serenityjs/item";
-import { CustomBlockType } from "@serenityjs/block";
 import { PlayerStatus } from "@serenityjs/world";
+import { CompoundTag, Tag } from "@serenityjs/nbt";
 
 import { ResourcePack } from "../resource-packs/resource-pack-manager";
 
@@ -123,10 +124,14 @@ class ResourcePackClientResponse extends SerenityHandler {
 				// Set the player as connected
 				player.status = PlayerStatus.Connected;
 
-				const blocks: Array<BlockProperties> = CustomBlockType.getAll().map(
-					(type) => {
+				// Get the player's world
+				const world = player.dimension.world;
+
+				const blocks: Array<BlockProperties> = world.blocks
+					.getAllCustomTypes()
+					.map((type) => {
 						// Get the item type from the block type
-						const item = ItemType.resolve(type) as CustomItemType;
+						const item = world.items.resolveType(type) as CustomItemType;
 
 						// Get the block nbt and item nbt
 						const blockNbt = type.nbt;
@@ -158,8 +163,7 @@ class ResourcePackClientResponse extends SerenityHandler {
 							name: item.identifier,
 							nbt: blockNbt
 						};
-					}
-				);
+					});
 
 				const packet = new StartGamePacket();
 				packet.entityId = player.unique;
@@ -434,9 +438,9 @@ class ResourcePackClientResponse extends SerenityHandler {
 				packet.blockProperties = blocks;
 
 				// Map the custom items to the packet
-				packet.items = ItemType.getAll().map((item) =>
-					ItemType.toItemData(item)
-				);
+				packet.items = world.items
+					.getAllTypes()
+					.map((item) => ItemType.toItemData(item));
 
 				packet.multiplayerCorrelationId = "<raknet>a555-7ece-2f1c-8f69";
 				packet.serverAuthoritativeInventory = true;
@@ -474,11 +478,31 @@ class ResourcePackClientResponse extends SerenityHandler {
 					};
 				});
 
+				// Map the entities to a nbt tag
+				const entities = world.entities.getAllTypes().map((type) => {
+					const tag = new CompoundTag();
+					tag.createStringTag("bid", "");
+					tag.createByteTag("hasspawnegg", 1);
+					tag.createStringTag("id", type.identifier);
+					tag.createIntTag("rid", type.network);
+					tag.createByteTag("summonable", 1);
+
+					return tag;
+				});
+
+				// Create root compound tag
+				const root = new CompoundTag();
+				root.createListTag("idlist", Tag.Compound, entities);
+
+				// Create the actors packet
+				const actors = new AvailableActorIdentifiersPacket();
+				actors.data = root;
+
 				const status = new PlayStatusPacket();
 				status.status = PlayStatus.PlayerSpawn;
 
 				// Send the spawn sequence
-				session.send(packet, biomes, content, status, crafting);
+				session.send(packet, biomes, content, status, crafting, actors);
 
 				// Add the player to the connecting map
 				this.serenity.connecting.set(player.session, [