+++ packages/world/src/chunk/chunk.ts
@@ -76,6 +76,9 @@ export class Chunk {
 	 */
 	public readonly subchunks: Array<SubChunk>;
 
+	/**
+	 * A cached buffer of the serialized chunk.
+	 */
 	public cache: Buffer | null = null;
 
 	/**
@@ -115,8 +118,9 @@ export class Chunk {
 	 * @param y The Y coordinate.
 	 * @param z The Z coordinate.
 	 */
-	public getPermutation(x: number, y: number, z: number): BlockPermutation {
+	public getPermutation(position: IPosition): BlockPermutation {
 		// Correct the Y level for the overworld.
+		const { x, y, z } = position;
 		const yf = this.type === DimensionType.Overworld ? y + 64 : y;
 
 		// Get the sub chunk.
@@ -137,12 +141,12 @@ export class Chunk {
 	 * @param permutation The permutation.
 	 */
 	public setPermutation(
-		x: number,
-		y: number,
-		z: number,
+		position: IPosition,
 		permutation: BlockPermutation,
 		dirty = true
 	): void {
+		const { x, y, z } = position;
+
 		// Correct the Y level for the overworld.
 		const yf = this.type === DimensionType.Overworld ? y + 64 : y;
 
@@ -171,7 +175,7 @@ export class Chunk {
 		// Get the Y level.
 		for (let y = position.y; y >= -64; y--) {
 			// Get the permutation at the position.
-			const permutation = this.getPermutation(position.x, y, position.z);
+			const permutation = this.getPermutation({ ...position, y });
 
 			// Check if the permutation is air or is not solid.
 			if (permutation.type.identifier === BlockIdentifier.Air) continue;
@@ -193,7 +197,7 @@ export class Chunk {
 	public getBottommostLevel(position: IPosition): number {
 		// Get the Y level.
 		for (let y = 0; y <= position.y; y++) {
-			const permutation = this.getPermutation(position.x, y, position.z);
+			const permutation = this.getPermutation({ ...position, y });
 			if (permutation.type.identifier !== BlockIdentifier.Air) return y;
 		}
 
@@ -224,7 +228,8 @@ export class Chunk {
 
 				// Create a new sub chunk.
 				const subchunk = new SubChunk();
-				subchunk.index = ndex - 4;
+				subchunk.index =
+					this.type === DimensionType.Overworld ? ndex - 4 : ndex;
 
 				// Set the sub chunk.
 				this.subchunks[ndex] = subchunk;