+++ packages/world/src/components/block/index.ts
@@ -12,3 +12,5 @@ export * from "./collision-box";
 export * from "./sign";
 export * from "./loot";
 export * from "./supported";
+export * from "./tillable";
+export * from "./gravity";