+++ packages/world/src/generator/overworld/index.ts
@@ -1,2 +1,2 @@
-export * from './generator';
-export * from './worker';
+export * from "./generator";
+export * from "./worker";