+++ packages/emitter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @serenityjs/emitter
 
+## 0.5.2
+
+### Patch Changes
+
+- 03a55ba: init v0.5.2-beta
+
+## 0.5.1
+
+### Patch Changes
+
+- 2880acc: init v0.5.1-beta
+
 ## 0.5.0
 
 ### Minor Changes